requirements.txt
flask
pytesseract
pdf2image
pillow
PyPDF2

